// src/helper/enums/events.ts
export enum EventsEnum {
  /** 执行部分更新Web命令完成 操作触发的事件 */
  UPDATE_WEB_MSG = 'update.web.msg',
  /** 执行部分更新Admin命令完成 操作触发的事件 */
  UPDATE_ADMIN_MSG = 'update.admin.msg',
  /** 执行部分更新Serve命令完成 操作触发的事件 */
  UPDATE_SERVE_MSG = 'update.serve.msg',
}